refactor: remove workspace-database-migration.ts
